Modifying RMG to read in seed gen from chemkin
Modfying RMG to read in chebychev polynomials of different
length from seed mechanism generated by existing chemkin
file.

Also removed the check of MRH to remove tacked on numbers
in reading in the seed mech file as this was causing conflicts
when we have two species of same formula just appended with
a number example C4H8(1) and C4H8(2) in the seed mechanism.
Warning this could affect restart and needs to be checked