Updated DASSL to handle up to five products in reactions.
The reaction arrays for regular, third body, Lindemann, and Troe reactions
needed to be updated to include another element for the new product. (Aside:
why aren't these arrays two-dimensional?)

DASPK has not yet been updated, since it would be even more involved, and
I'm not set up to test it at the moment. However, the sorts of changes made
here should be similar to changes that would be needed there.

Note that the maximum number of reactants and products is still hardcoded in
many places.

Mimics changes done by @jwallen in commit
affe5956095d5b16fcb46051018522a7ef8fe98e

I have tested the minimal example after this commit and it gives exactly the
same results as before changing the product number for the ODE
so I expect it to be correct, however it would be nice to check against
chemkin.